+++ src/main/java/kr/codesquad/ladder/domain/InvalidNameOfPeopleException.java
@@ -1,4 +1,4 @@
-package kr.codesquad.domain;
+package kr.codesquad.ladder.domain;
 
 public class InvalidNameOfPeopleException extends RuntimeException {
