+++ src/main/java/kr/codesquad/ladder/domain/LadderPartGenerator.java
@@ -1,4 +1,4 @@
-package kr.codesquad.domain;
+package kr.codesquad.ladder.domain;
 
 import java.util.ArrayList;
 import java.util.List;
