+++ src/main/java/kr/codesquad/ladder/domain/LadderValidator.java
@@ -1,4 +1,4 @@
-package kr.codesquad.domain;
+package kr.codesquad.ladder.domain;
 
 import java.util.Arrays;
 import java.util.regex.Pattern;
