+++ src/main/java/kr/codesquad/ladder/view/LadderReader.java
@@ -1,4 +1,4 @@
-package kr.codesquad.view;
+package kr.codesquad.ladder.view;
 
 import java.util.List;
