+++ src/test/java/kr/codesquad/LadderConsoleReaderTest.java
@@ -7,11 +7,11 @@ import java.io.InputStreamReader;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
-import kr.codesquad.domain.LadderValidator;
-import kr.codesquad.view.LadderConsoleReader;
-import kr.codesquad.view.LadderConsoleWriter;
-import kr.codesquad.view.LadderReader;
-import kr.codesquad.view.LadderWriter;
+import kr.codesquad.ladder.domain.LadderValidator;
+import kr.codesquad.ladder.view.LadderConsoleReader;
+import kr.codesquad.ladder.view.LadderConsoleWriter;
+import kr.codesquad.ladder.view.LadderReader;
+import kr.codesquad.ladder.view.LadderWriter;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
